+++ extensions/data-plane-selector/selector-store/src/main/java/org/eclipse/dataspaceconnector/dataplane/selector/store/DataPlaneInstanceStoreExtension.java
@@ -1,0 +1,28 @@
+/*
+ *  Copyright (c) 2020 - 2022 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - initial API and implementation
+ *
+ */
+
+package org.eclipse.dataspaceconnector.dataplane.selector.store;
+
+import org.eclipse.dataspaceconnector.spi.system.Provides;
+import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
+import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
+
+@Provides(DataPlaneInstanceStore.class)
+public class DataPlaneInstanceStoreExtension implements ServiceExtension {
+    
+    @Override
+    public void initialize(ServiceExtensionContext context) {
+        context.registerService(DataPlaneInstanceStore.class, new DefaultDataPlaneInstanceStore());
+    }
+}