+++ extensions/data-plane/data-plane-http/src/test/java/org/eclipse/dataspaceconnector/dataplane/http/pipeline/HttpDataSourceFactoryTest.java
@@ -9,13 +9,17 @@
  *
  *  Contributors:
  *       Microsoft Corporation - initial API and implementation
+ *       Amadeus - test retrieval of auth code from vault
+ *       Amadeus - add test for mapping of path segments
  *
  */
+
 package org.eclipse.dataspaceconnector.dataplane.http.pipeline;
 
 import net.jodah.failsafe.RetryPolicy;
 import okhttp3.MediaType;
 import okhttp3.OkHttpClient;
+import org.eclipse.dataspaceconnector.dataplane.http.HttpTestFixtures;
 import org.eclipse.dataspaceconnector.spi.EdcException;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.eclipse.dataspaceconnector.spi.security.Vault;
@@ -31,14 +35,14 @@ import org.junit.jupiter.params.provider.ValueSource;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.UUID;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.eclipse.dataspaceconnector.dataplane.http.HttpTestFixtures.createDataAddress;
-import static org.eclipse.dataspaceconnector.dataplane.http.HttpTestFixtures.createRequest;
 import static org.eclipse.dataspaceconnector.dataplane.spi.schema.DataFlowRequestSchema.BODY;
 import static org.eclipse.dataspaceconnector.dataplane.spi.schema.DataFlowRequestSchema.MEDIA_TYPE;
 import static org.eclipse.dataspaceconnector.dataplane.spi.schema.DataFlowRequestSchema.METHOD;
+import static org.eclipse.dataspaceconnector.dataplane.spi.schema.DataFlowRequestSchema.PATH;
 import static org.eclipse.dataspaceconnector.dataplane.spi.schema.DataFlowRequestSchema.QUERY_PARAMS;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.AUTHENTICATION_CODE;
 import static org.eclipse.dataspaceconnector.spi.types.domain.http.HttpDataAddressSchema.AUTHENTICATION_KEY;
@@ -72,21 +76,30 @@ class HttpDataSourceFactoryTest {
     @EmptySource
     @ValueSource(strings = "dummy")
     void verifyCannotHandle(String type) {
-        assertThat(factory.canHandle(createRequest(type).build())).isFalse();
+        assertThat(factory.canHandle(HttpTestFixtures.createRequest(type).build())).isFalse();
     }
 
     @ParameterizedTest(name = "{index} {0}")
     @MethodSource("provideInvalidRequests")
-    void verifySourceValidationAndCreation_failure(String name, DataFlowRequest request) {
+    void verifySourceValidationAndCreation_failure(String name, TestInstance testInstance) {
+        var request = testInstance.createRequest();
+
         assertThat(factory.validate(request).failed()).isTrue();
         assertThrows(EdcException.class, () -> factory.createSource(request));
     }
 
     @ParameterizedTest(name = "{index} {0}")
-    @MethodSource("provideValidRequestsWithExpectedSource")
-    void verifySourceValidationAndCreation(String name, DataFlowRequest request, HttpDataSource expectedSource) {
+    @MethodSource("provideTestInstances")
+    void verifySourceValidationAndCreation(String name, TestInstance testInstance) {
         when(vaultMock.resolveSecret(TEST_SECRET_NAME)).thenReturn(TEST_SECRET_VALUE);
+
+        var request = testInstance.createRequest();
+        var expectedSource = testInstance.createDataSource();
+
         assertThat(factory.canHandle(request)).isTrue();
+
+        var test = factory.validate(request);
+
         assertThat(factory.validate(request).succeeded()).isTrue();
         var source = factory.createSource(request);
         assertThat(source).isNotNull();
@@ -97,7 +110,7 @@ class HttpDataSourceFactoryTest {
             try {
                 assertThat(f.get(source)).isEqualTo(f.get(expectedSource));
             } catch (IllegalAccessException e) {
-                throw new AssertionError(e);
+                throw new AssertionError("Comparison failed for field: " + f.getName());
             }
         });
     }
@@ -107,85 +120,200 @@ class HttpDataSourceFactoryTest {
      */
     private static Stream<Arguments> provideInvalidRequests() {
         var endpoint = "http://example.com";
-        var validAddress = createDataAddress(TYPE, Map.of(ENDPOINT, endpoint)).build();
-        var missingEndpoint = createDataAddress(TYPE, Map.of()).build();
+        var authKey = "apikey-test";
+
+        var missingMethod = TestInstance.newInstance()
+                .endpoint(endpoint);
+
+        var missingEndpoint = TestInstance.newInstance()
+                .method("GET");
+
+        var incompleteHeader = TestInstance.newInstance()
+                .method("GET")
+                .authKey(authKey)
+                .endpoint(endpoint);
+
+        var unknownMediaType = TestInstance.newInstance()
+                .method("POST")
+                .body("dummy", "hello world!")
+                .endpoint(endpoint);
+
         return Stream.of(
-                Arguments.of("MISSING METHOD", createRequest(Map.of(), validAddress, validAddress).build()),
-                Arguments.of("MISSING ENDPOINT", createHttpRequest(missingEndpoint, "GET")),
-                Arguments.of("UNHANDLED MEDIA TYPE", createHttpRequest(validAddress, "GET", Map.of(MEDIA_TYPE, "dummy", BODY, "body-test")))
+                Arguments.of("MISSING METHOD", missingMethod),
+                Arguments.of("MISSING ENDPOINT", missingEndpoint),
+                Arguments.of("INCOMPLETE HEADER", incompleteHeader),
+                Arguments.of("UNHANDLED MEDIA TYPE", unknownMediaType)
         );
     }
 
     /**
      * Serves some valid {@link DataFlowRequest} with the associated expected {@link HttpDataSource} that must be generated.
      */
-    private static Stream<Arguments> provideValidRequestsWithExpectedSource() {
+    private static Stream<Arguments> provideTestInstances() {
         var endpoint = "http://example.com";
         var name = "foo.json";
         var authKey = "apikey-test";
         var mediaType = "application/json";
         var body = "test";
         var queryParams = "?foo=bar";
-        var defaultDataAddress = createDataAddress(TYPE, Map.of(ENDPOINT, endpoint)).build();
-        var dataAddressWithName = createDataAddress(TYPE, Map.of(ENDPOINT, endpoint, NAME, name)).build();
-        var dataAddressWithHardcodedAuth = createDataAddress(TYPE, Map.of(ENDPOINT, endpoint, AUTHENTICATION_KEY, authKey, AUTHENTICATION_CODE, TEST_SECRET_VALUE)).build();
-        var dataAddressWithSecretBasedAuth = createDataAddress(TYPE, Map.of(ENDPOINT, endpoint, AUTHENTICATION_KEY, authKey, SECRET_NAME, TEST_SECRET_NAME)).build();
-        var dataAddressWithSecretNameOnly = createDataAddress(TYPE, Map.of(ENDPOINT, endpoint, SECRET_NAME, TEST_SECRET_NAME)).build();
 
-        var basicRequest = createHttpRequest(defaultDataAddress, "GET");
-        var expectedBasicSource = defaultHttpSource(basicRequest).build();
+        var get = TestInstance.newInstance()
+                .method("GET")
+                .endpoint(endpoint);
 
-        var requestWithName = createHttpRequest(dataAddressWithName, "GET");
-        var sourceWithName = defaultHttpSource(requestWithName).name(name).build();
+        var getWithPath = TestInstance.newInstance()
+                .method("GET")
+                .basePath("hello/world")
+                .endpoint(endpoint);
 
-        var requestWithQueryParams = createHttpRequest(defaultDataAddress, "GET", Map.of(QUERY_PARAMS, queryParams));
-        var sourceWithQueryParams = defaultHttpSource(requestWithQueryParams).queryParams(queryParams).build();
+        var getWithName = TestInstance.newInstance()
+                .method("GET")
+                .name(name)
+                .endpoint(endpoint);
 
-        var requestWithSecretBasedAuth = createHttpRequest(dataAddressWithSecretBasedAuth, "GET");
-        var sourceWithAuthHeaderFromSecret = defaultHttpSource(requestWithSecretBasedAuth).header(authKey, TEST_SECRET_VALUE).build();
+        var getWithQueryParams = TestInstance.newInstance()
+                .method("GET")
+                .queryParams(queryParams)
+                .endpoint(endpoint);
 
-        var requestWithAuthHardcoded = createHttpRequest(dataAddressWithHardcodedAuth, "GET");
-        var sourceWithAuthHeaderFromHardcode = defaultHttpSource(requestWithAuthHardcoded).header(authKey, TEST_SECRET_VALUE).build();
+        var getWithSecret = TestInstance.newInstance()
+                .method("GET")
+                .authHeader(authKey, TEST_SECRET_NAME, TEST_SECRET_VALUE)
+                .endpoint(endpoint);
 
-        var requestWithBody = createHttpRequest(defaultDataAddress, "POST", Map.of(MEDIA_TYPE, mediaType, BODY, body));
-        var sourceWithBody = defaultHttpSource(requestWithBody).requestBody(MediaType.get(mediaType), body).build();
+        var getWithAuthCode = TestInstance.newInstance()
+                .method("GET")
+                .authHeader(authKey, TEST_SECRET_VALUE)
+                .endpoint(endpoint);
 
-        var requestWithIncompleteHeader = createHttpRequest(dataAddressWithSecretNameOnly, "GET");
-        var expectedSourceWithoutHeader = defaultHttpSource(requestWithIncompleteHeader).build();
+        var post = TestInstance.newInstance()
+                .method("POST")
+                .body(mediaType, body)
+                .endpoint(endpoint);
 
-        var requestWithMediaTypeOnly = createHttpRequest(defaultDataAddress, "PUT", Map.of(ENDPOINT, endpoint, BODY, body));
-        var expectedSourceWithoutBody = defaultHttpSource(requestWithMediaTypeOnly).build();
+
+        var ignoreBodyWithoutMediaType = TestInstance.newInstance()
+                .method("POST")
+                .body(body)
+                .endpoint(endpoint);
 
         return Stream.of(
-                Arguments.of("BASIC SOURCE", basicRequest, expectedBasicSource),
-                Arguments.of("WITH NAME", requestWithName, sourceWithName),
-                Arguments.of("WITH QUERY PARAMS", requestWithQueryParams, sourceWithQueryParams),
-                Arguments.of("WITH AUTH SECRET", requestWithSecretBasedAuth, sourceWithAuthHeaderFromSecret),
-                Arguments.of("WITH AUTH HARDCODED", requestWithAuthHardcoded, sourceWithAuthHeaderFromHardcode),
-                Arguments.of("WITH BODY", requestWithBody, sourceWithBody),
-                Arguments.of("WITHOUT AUTH KEY", requestWithIncompleteHeader, expectedSourceWithoutHeader),
-                Arguments.of("WITHOUT MEDIA TYPE", requestWithMediaTypeOnly, expectedSourceWithoutBody)
+                Arguments.of("GET", get),
+                Arguments.of("GET WITH NAME", getWithName),
+                Arguments.of("GET WITH PATH", getWithPath),
+                Arguments.of("GET WITH QUERY PARAMS", getWithQueryParams),
+                Arguments.of("GET WITH SECRET", getWithSecret),
+                Arguments.of("WITH AUTH CODE", getWithAuthCode),
+                Arguments.of("POST", post),
+                Arguments.of("IGNORE BODY WITHOUT MEDIA TYPE", ignoreBodyWithoutMediaType)
         );
     }
 
-    private static DataFlowRequest createHttpRequest(DataAddress address, String method) {
-        return createHttpRequest(address, method, Map.of());
-    }
+    /**
+     * One test instance for parameterized tests.
+     */
+    private static final class TestInstance {
 
-    private static DataFlowRequest createHttpRequest(DataAddress address, String method, Map<String, String> additional) {
-        var properties = new HashMap<String, String>();
-        properties.put(METHOD, method);
-        properties.putAll(additional);
-        return createRequest(TYPE).sourceDataAddress(address).properties(properties).build();
-    }
+        private static final DataAddress DUMMY_ADDRESS = DataAddress.Builder.newInstance()
+                .type("dummy")
+                .build();
 
-    private static HttpDataSource.Builder defaultHttpSource(DataFlowRequest request) {
-        return HttpDataSource.Builder.newInstance()
-                .sourceUrl(request.getSourceDataAddress().getProperty(ENDPOINT))
-                .method(request.getProperties().get(METHOD))
-                .requestId(request.getId())
-                .httpClient(HTTP_CLIENT)
-                .monitor(MONITOR)
-                .retryPolicy(RETRY_POLICY);
+        private final Map<String, String> props;
+        private final DataAddress.Builder address;
+        private final DataFlowRequest.Builder request;
+        private final HttpDataSource.Builder source;
+
+        private TestInstance() {
+            var requestId = UUID.randomUUID().toString();
+            props = new HashMap<>();
+            address = DataAddress.Builder.newInstance().type(TYPE);
+            request = DataFlowRequest.Builder.newInstance().processId("1").id(requestId);
+            source = defaultHttpSource().requestId(requestId);
+        }
+
+        public static TestInstance newInstance() {
+            return new TestInstance();
+        }
+
+        public DataFlowRequest createRequest() {
+            return request.sourceDataAddress(address.build()).destinationDataAddress(DUMMY_ADDRESS).properties(props).build();
+        }
+
+        public HttpDataSource createDataSource() {
+            return source.build();
+        }
+
+        public TestInstance endpoint(String endpoint) {
+            this.address.property(ENDPOINT, endpoint);
+            source.sourceUrl(endpoint);
+            return this;
+        }
+
+        public TestInstance method(String method) {
+            props.put(METHOD, method);
+            source.method(method);
+            return this;
+        }
+
+        public TestInstance name(String name) {
+            this.address.property(NAME, name);
+            source.name(name);
+            return this;
+        }
+
+        public TestInstance authKey(String authKey) {
+            this.address.property(AUTHENTICATION_KEY, authKey);
+            return this;
+        }
+
+        public TestInstance authCode(String authCode) {
+            this.address.property(AUTHENTICATION_CODE, authCode);
+            return this;
+        }
+
+        public TestInstance authHeader(String key, String code) {
+            this.address.property(AUTHENTICATION_KEY, key);
+            this.address.property(AUTHENTICATION_CODE, code);
+            source.header(key, code);
+            return this;
+        }
+
+        public TestInstance authHeader(String key, String secret, String code) {
+            this.address.property(AUTHENTICATION_KEY, key);
+            this.address.property(SECRET_NAME, secret);
+            source.header(key, code);
+            return this;
+        }
+
+        public TestInstance body(String body) {
+            props.put(BODY, body);
+            return this;
+        }
+
+        public TestInstance body(String mediaType, String body) {
+            props.put(MEDIA_TYPE, mediaType);
+            props.put(BODY, body);
+            source.requestBody(MediaType.parse(mediaType), body);
+            return this;
+        }
+
+        public TestInstance basePath(String basePath) {
+            props.put(PATH, "hello/world");
+            source.name(basePath);
+            return this;
+        }
+
+        public TestInstance queryParams(String queryParams) {
+            props.put(QUERY_PARAMS, queryParams);
+            source.queryParams(queryParams);
+            return this;
+        }
+
+        private HttpDataSource.Builder defaultHttpSource() {
+            return HttpDataSource.Builder.newInstance()
+                    .httpClient(HTTP_CLIENT)
+                    .monitor(MONITOR)
+                    .retryPolicy(RETRY_POLICY);
+        }
     }
 }