+++ extensions/in-memory/assetindex-memory/src/test/java/org/eclipse/dataspaceconnector/metadata/memory/InMemoryAssetLoaderIndexTest.java
@@ -82,7 +82,7 @@ public class InMemoryAssetLoaderIndexTest {
 
     @BeforeEach
     void setup() {
-        assetLoader = new InMemoryAssetIndex(new AssetPredicateConverter());
+        assetLoader = new InMemoryAssetIndex();
     }
 
     private Asset createAsset(String name, String id) {