+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+
 # Changelog
 
 All notable changes to this project will be documented in this file - formatted and maintained according to the rules
@@ -13,6 +14,7 @@ the detailed section referring to by linking pull requests or issues.
 
 * Removed deprecated code
 * Improved CosmosDB interaction
+* Added DPF Selector
 * Updated Data Management API
 
 ### Detailed Changes
@@ -27,6 +29,11 @@ the detailed section referring to by linking pull requests or issues.
 * Implement S3BucketReader (#675)
 * Add configuration setting for state machine batch size (#872)
 * Add Jetty context alias for IDS API (#815)
+* Implement Asset service for Data Management API (#931)
+* Added embedded and remote DPF Selector (#832)
+* Pass path information into http data source through DPF public API (#929)
+* Add instructions for observability sample with Azure Application Insights (#928)
+* Add interface `WebServer` to `web-spi` (#921)
 * New Policy Extension for Token Claim Validation (#953)
 
 #### Changed
@@ -39,8 +46,11 @@ the detailed section referring to by linking pull requests or issues.
 * Enable pluggable transfer service in DPF (#844)
 * Apply 2-state transition pattern to `ContractNegotiationManager`s (#870)
 * Apply 2-state transition pattern to `TransferProcessManager` (#831)
+* Update build system to Java 17 (#934)
 * Refactor (=generify) transformer subsystem (#779)
 * Extract interfaces for every api controller class to improve swagger documentation (#891)
+* Instrument executors with metrics (#912)
+* Call the listeners before the state transition is persisted. (#876)
 
 #### Removed
 
@@ -50,8 +60,10 @@ the detailed section referring to by linking pull requests or issues.
 * Remove module `:samples:other:streaming` (#889)
 
 #### Fixed
+
 * Flaky S3 StatusChecker Test (#794)
-* Added missing Data Management Asset controller openapi (#853) 
+* Added missing Data Management Asset controller openapi (#853)
+* Policy deserialization (#898)
 
 ---